Getting Started with FVMFramework

This tutorial will teach you how to use the tools provided by FVMFramework to solve a 3D heat transfer problem

In this tutorial, we will be solving the following heat transfer equation:

\[Q_{a, in} = -k A \frac{(T_{a}-T_{b})}{\Delta d}\]

where:

\[Q_{a,in}\]
is the heat flux going into cell a in [W]

\[k\]
is the thermal conductivity of the material in [W/(m*K)]

\[A\]
is the area of the face shared by boths cells in [m^2]

\[T_{a}\]
and $T_{b}$ are the temperatures of cell a and cell b respectively in [K]

\[d\]
is the distance between the cell centers of each cell in [m]

FVMFramework.create_fvm_config — Function

create_fvm_config(grid, u_proto)

Create a SimulationConfigInfo struct with the given grid and initial conditions.

Arguments

grid::Ferrite.Grid: The grid to use for the simulation.

u_proto::ComponentVector: The initial conditions for the simulation.

Returns

SimulationConfigInfo: The created simulation config struct.

using Revise

using FVMFramework

using Ferrite
using FerriteGmsh
using OrdinaryDiffEq
using SparseArrays
using ComponentArrays
using NonlinearSolve
import SparseConnectivityTracer, ADTypes
using ILUZero
using StaticArrays

using Unitful

The general workflow of this package is to:

Define a simple mesh using Ferrite.jl or import a mesh using FerriteGmsh.jl

Create a config struct

add cached fields, special caches, and optimized parameters

add regions (groups of volumes with similar properties) and patches (groups of faces with similar properties)

define the physics functions including physics functions for calculating fluxes between faces and fluxes between

grid_dimensions = (10, 10, 10)
left = Ferrite.Vec{3}((0.0, 0.0, 0.0))
right = Ferrite.Vec{3}((1.0, 1.0, 1.0))
grid = generate_grid(Tetrahedron, grid_dimensions, left, right)

cell_half_dist = (right[1] / grid_dimensions[1]) / 2
addcellset!(grid, "copper", x -> x[1] <= (left[1] + (right[1] / 2) + cell_half_dist))
addcellset!(grid, "steel", x -> x[1] >= left[1] + (right[1] / 2))

julia

u_proto = ComponentVector(
    temp = zeros(length(grid.cells))u"K",
)

config = createfvmconfig(grid, u_proto)

nfaces = length(config.geo.cellneighbor_areas[1])

struct Solid <: AbstractPhysics end

#property updating/retrieval

#variable summation

#internal physics

#sources

#boundary conditions

#capacities

addsetupsyms!( config; cachesymsandunits = (heat = u"J",), specialcaches = ComponentVector(), secondordersyms = [], optimized_parameters = ComponentVector() )

function capheatfluxtotempchange!(du, u, cellid, vol) # J/s /= m^3 * kgm^3 * J/(kgK) # = K/s #@show du.heat[cellid] #@show vol #@show u.rho[cellid] #@show u.cp[cellid] du.temp[cellid] += du.heat[cellid] / (vol * u.rho[cellid] * u.cp[cellid]) #@show du.temp[cellid] end

function capmassfluxtopressurechange!(du, u, cellid, vol) # kg/s /= (m^3 / (J/(molK) * K)) #remember: J = Pam^3 # = Pa/s dumoles = du.mass[cellid] / u.mwavg[cellid] du.pressure[cellid] += (dumoles * u.Rgas[cellid] * u.temp[cell_id]) / vol end

function capspeciesmassfluxtomassfractionchange!(du, u, cellid, vol) totalmass = vol * u.rho[cellid]

for_fields!(du.mass_fractions, u.mass_fractions, du.species_mass_flows) do species, du_mass_fractions, u_mass_fractions, species_mass_flows
    du_mass_fractions[species[cell_id]] += (species_mass_flows[species[cell_id]] - u_mass_fractions[species[cell_id]] * du.mass[cell_id]) / total_mass
end

end

addregion!( config, "copper"; type = Solid(), initialconditions = ComponentVector( temp = 270.0u"°C", ), properties = ComponentVector( k = 237.0u"W/(mK)", rho = 2700.0u"kg/m^3", cp = 921.0u"J/(kgK)", ), propertyupdatefunction = function copperpropertyupdate!(properties, u)

end,
region_function =
function heat_transfer!(du, u, cell_id, vol)
    cap_heat_flux_to_temp_change!(du, u, cell_id, vol)
end

)

addregion!( config, "steel"; type = Solid(), initialconditions = ComponentVector( temp = 350.0u"°C", ), properties = ComponentVector( k = 123.0u"W/(mK)", rho = 7800.0u"kg/m^3", cp = 450.0u"J/(kgK)", ), propertyupdatefunction = function steelpropertyupdate!(properties, u)

end,
region_function =
function heat_transfer!(du, u, p, t, cell_id, vol)
    cap_heat_flux_to_temp_change!(du, u, cell_id, vol)
end

)

function heatdiffusion!( du, u, p, t, idxa, idxb, faceidx, area, norm, dist, vola, volb ) keffective = 2 * u.k[idxa] * u.k[idxb] / (u.k[idxa] + u.k[idx_b])

grad_T = (u.temp[idx_b] - u.temp[idx_a]) / dist

du.heat[idx_a] -= -k_effective * grad_T * area

end

function solidsolidflux!( du, u, p, t, idxa, idxb, faceidx, cellneighborareas, cellneighbornormals, cellneighbordistances, cellvolumes )

#hmm, perhaps these physics functions need to be more strictly typed
#Checking profview, I'm getting some runtime dispatch and GC here, I don't know why 
heat_diffusion!(
    du, u, p, t,
    idx_a, idx_b, face_idx,
    cell_neighbor_areas[idx_a][face_idx], cell_neighbor_normals[idx_a][face_idx], cell_neighbor_distances[idx_a][face_idx],
    cell_volumes[idx_a], cell_volumes[idx_b]
)

end

function connectionmapfunction(typea, typeb) typeof(typea) <: Solid && typeof(typeb) <: Solid && return solidsolidflux! end

du0vec, u0vec, geo, system = finishfvmconfig(config, connectionmapfunction, check_units = false);

function solvesystem!(du, u, p, t, geo, system) solveconnectiongroups!(du, u, p, t, geo, system) solvecontrollergroups!(du, u, p, t, geo, system) solvepatchgroups!(du, u, p, t, geo, system) solveregion_groups!(du, u, p, t, geo, system) end

fclosureimplicit = (du, u, p, t) -> fvmoperator!(du, u, p, t, solvesystem!, geo, system)

p_guess = 0.0

testprob = ODEProblem(fclosureimplicit, u0vec, (0.0, 1000.0), pguess) sol = solve(testprob, Tsit5(), tspan = (0.0, 10.0))

t0 = 0.0 tMax = 1000.0 tspan = (t0, tMax)

detector = SparseConnectivityTracer.TracerLocalSparsityDetector() #not sure if pure TracerSparsityDetector is faster

jacsparsity = ADTypes.jacobiansparsity( (du, u) -> fclosureimplicit(du, u, pguess, 0.0), du0vec, u0_vec, detector )

odefunc = ODEFunction(fclosureimplicit, jacprototype = float.(jac_sparsity))

implicitprob = ODEProblem(odefunc, u0vec, tspan, pguess)

desiredsteps = 100 saveinterval = (tspan[end] / desired_steps)

@time sol = solve(implicitprob, FBDF(linsolve = KrylovJLGMRES(), precs = iluzero, concretejac = true), callback = approximatetimetofinishcb) @time sol = solve(implicitprob, FBDF(linsolve = KLUFactorization(), precs = iluzero, concretejac = true), callback = approximatetimetofinishcb) @time sol = solve(implicitprob, FBDF(precs = iluzero, concretejac = true), callback = approximatetimetofinish_cb) #728.605 ms (341178 allocations: 1.08 GiB) (non-multithreaded) #787.708 ms (202070 allocations: 1.07 GiB) (only connections multithreading) #799.862 ms (219386 allocations: 1.07 GiB) (everything multithreaded) #864.517 ms (211023 allocations: 1.07 GiB) (only regions multithreading)

VSCodeServer.@profview sol = solve(implicitprob, FBDF(linsolve = KrylovJLGMRES(), precs = iluzero, concretejac = true), callback = approximatetimetofinish_cb) #algebraicmultigrid is only better for more than 1e6 cells

record_sol = true

sim_file = @FILE

dunamed, unamed = regeneratefvmstate(sol, system, solvesystem!, geo, pguess)

if recordsol == true soltovtk(sol, unamed, grid, sim_file) end